% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmixPenalized.R
\docType{methods}
\name{flexmix,formula,ANY,ANY,ANY,FLXMCLsvm-method}
\alias{flexmix,formula,ANY,ANY,ANY,FLXMCLmultinom-method}
\alias{flexmix,formula,ANY,ANY,ANY,FLXMCLnnet-method}
\alias{flexmix,formula,ANY,ANY,ANY,FLXMCLsvm-method}
\alias{flexmixPenalized}
\alias{flexmixPenalized,formula,ANY,ANY,ANY,list-method}
\title{Mixtures of Penalized Classification Methods}
\usage{
\S4method{flexmix}{formula,ANY,ANY,ANY,FLXMCLsvm}(formula, data = list(),
  k = NULL, cluster = NULL, model = NULL, concomitant = NULL,
  control = NULL, weights = NULL)

\S4method{flexmix}{formula,ANY,ANY,ANY,FLXMCLmultinom}(formula, data = list(),
  k = NULL, cluster = NULL, model = NULL, concomitant = NULL,
  control = NULL, weights = NULL)

\S4method{flexmix}{formula,ANY,ANY,ANY,FLXMCLnnet}(formula, data = list(),
  k = NULL, cluster = NULL, model = NULL, concomitant = NULL,
  control = NULL, weights = NULL)

flexmixPenalized(formula, data = list(), k = NULL, cluster = NULL,
  model = NULL, concomitant = NULL, control = NULL, weights = NULL)

\S4method{flexmixPenalized}{formula,ANY,ANY,ANY,list}(formula, data = list(),
  k = NULL, cluster = NULL, model = NULL, concomitant = NULL,
  control = NULL, weights = NULL)
}
\description{
Extension of function \code{\link[flexmix]{flexmix}} in package \pkg{flexmix} that allows to fit mixtures of penalized classifiers. 
Not to be called by the user.
}

