% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmixPenalized.R
\docType{methods}
\name{FLXdeterminePostunscaled,FLXMCLsvm-method}
\alias{FLXdeterminePostunscaled,FLXMCLmultinom-method}
\alias{FLXdeterminePostunscaled,FLXMCLnnet-method}
\alias{FLXdeterminePostunscaled,FLXMCLsvm-method}
\title{Internal Functions}
\usage{
\S4method{FLXdeterminePostunscaled}{FLXMCLsvm}(model, components, ...)

\S4method{FLXdeterminePostunscaled}{FLXMCLmultinom}(model, components, ...)

\S4method{FLXdeterminePostunscaled}{FLXMCLnnet}(model, components, ...)
}
\description{
Internal functions not to be called by the user.
}

