% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexmixPenalized.R
\docType{methods}
\name{FLXfitPenalized}
\alias{FLXfitPenalized}
\alias{FLXfitPenalized,list-method}
\title{Mixtures of Penalized Classifiers}
\usage{
FLXfitPenalized(model, concomitant, control, postunscaled = NULL, groups,
  weights)

\S4method{FLXfitPenalized}{list}(model, concomitant, control,
  postunscaled = NULL, groups, weights)
}
\description{
Extension of function \code{\link[flexmix]{FLXfit}} in package \pkg{flexmix} that allows to fit mixtures of penalized classifiers.
Not to be called by the user.
}

