% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLXMCLmajority.R
\docType{methods}
\name{FLXMCLmajority}
\alias{FLXMCLmajority}
\alias{FLXgetModelmatrix,FLXMCLmajority-method}
\title{Mixtures of Majority Classifiers}
\usage{
FLXMCLmajority(formula = . ~ ., ...)

\S4method{FLXgetModelmatrix}{FLXMCLmajority}(model, data, formula, lhs = TRUE,
  ...)
}
\arguments{
\item{formula}{A formula which is interpreted relative to the formula specified in the call to \code{\link[flexmix]{flexmix}} using \code{\link[stats]{update.formula}}. 
Only the left-hand side (response) of the formula is used. Default is to use the original \code{\link[flexmix]{flexmix}} model formula.}

\item{\dots}{Further arguments to and from other methods.}
}
\value{
Returns an object of class \code{FLXMCLmajority} inheriting from \code{FLXMCL}.
}
\description{
This is a model driver for \code{\link[flexmix]{flexmix}} from package \pkg{flexmix} implementing mixtures of majority classifiers.
}
\examples{
library(benchData)
data <- flashData(1000)
data$x <- scale(data$x)
grid <- expand.grid(x.1=seq(-6,6,0.2), x.2=seq(-4,4,0.2))

cluster <- kmeans(data$x, center = 4)$cluster
model <- FLXMCLmajority()
fit <- flexmix(y ~ ., data = as.data.frame(data), concomitant = FLXPmultinom(~ x.1 + x.2), model = model, cluster = cluster)

## prediction with aggregation depending on membership in mixture components
pred.grid <- mypredict(fit, newdata = grid, aggregate = TRUE)
image(seq(-6,6,0.2), seq(-4,4,0.2), matrix(pred.grid[[1]][,1], length(seq(-6,6,0.2))))
contour(seq(-6,6,0.2), seq(-4,4,0.2), matrix(pred.grid[[1]][,1], length(seq(-6,6,0.2))), add  = TRUE)
points(data$x, pch = as.character(data$y))

## local membership
loc.grid <- prior(fit, newdata = grid)
contour(seq(-6,6,0.2), seq(-4,4,0.2), matrix(loc.grid[,1], length(seq(-6,6,0.2))), add  = TRUE)
contour(seq(-6,6,0.2), seq(-4,4,0.2), matrix(loc.grid[,2], length(seq(-6,6,0.2))), add  = TRUE)
contour(seq(-6,6,0.2), seq(-4,4,0.2), matrix(loc.grid[,3], length(seq(-6,6,0.2))), add  = TRUE)
contour(seq(-6,6,0.2), seq(-4,4,0.2), matrix(loc.grid[,4], length(seq(-6,6,0.2))), add  = TRUE)
}
\seealso{
Other mixtures majority: \code{\link{FLXMCL-class}},
  \code{\link{FLXMCLmajorityGenerative}}

Other mixtures majority: \code{\link{FLXMCL-class}},
  \code{\link{FLXMCLmajorityGenerative}}
}

