% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLXMCLsvm.R
\docType{methods}
\name{FLXMCLsvm}
\alias{FLXMCLsvm}
\alias{FLXgetModelmatrix,FLXMCLsvm-method}
\title{Mixtures of Support Vector Machines}
\usage{
FLXMCLsvm(formula = . ~ ., ...)

\S4method{FLXgetModelmatrix}{FLXMCLsvm}(model, data, formula, lhs = TRUE, ...)
}
\arguments{
\item{formula}{A formula which is interpreted relative to the formula specified in the call to \code{\link[flexmix]{flexmix}} 
   using \code{\link[stats]{update.formula}}. 
Only the left-hand side (response) of the formula is used. Default is to use the original \code{\link[flexmix]{flexmix}} model formula.}

\item{\dots}{Further arguments to and from other methods, especially to \code{\link{wsvm}}.}
}
\value{
Returns an object of class \code{FLXMCLsvm} inheriting from \code{FLXMCL}.
}
\description{
This is a model driver for \code{\link[flexmix]{flexmix}} implementing mixtures of Support Vector Machines for classification.
}
\examples{
library(benchData)
data <- flashData(1000)
data$x <- scale(data$x)
grid <- expand.grid(x.1 = seq(-6,6,0.2), x.2 = seq(-4,4,0.2))

cluster <- kmeans(data$x, center = 2)$cluster
model <- FLXMCLsvm(kernel = "linear", fitted = FALSE)
fit <- flexmix(y ~ ., data = as.data.frame(data), concomitant = FLXPmultinom(~ x.1 + x.2), model = model, cluster = cluster)

## prediction for single component models without aggregation
pred.grid <- predict(fit, newdata = grid)
image(seq(-6,6,0.2), seq(-4,4,0.2), matrix(pred.grid[[1]][,1], length(seq(-6,6,0.2))))
contour(seq(-6,6,0.2), seq(-4,4,0.2), matrix(pred.grid[[1]][,1], length(seq(-6,6,0.2))), add = TRUE)
points(data$x, pch = as.character(data$y))

image(seq(-6,6,0.2), seq(-4,4,0.2), matrix(pred.grid[[2]][,1], length(seq(-6,6,0.2))))
contour(seq(-6,6,0.2), seq(-4,4,0.2), matrix(pred.grid[[2]][,1], length(seq(-6,6,0.2))), add = TRUE)
points(data$x, pch = as.character(data$y))

## prediction with aggregation depending on membership in mixture components
pred.grid <- mypredict(fit, newdata = grid, aggregate = TRUE)
image(seq(-6,6,0.2), seq(-4,4,0.2), matrix(pred.grid[[1]][,1], length(seq(-6,6,0.2))))
contour(seq(-6,6,0.2), seq(-4,4,0.2), matrix(pred.grid[[1]][,1], length(seq(-6,6,0.2))), add  = TRUE)
points(data$x, pch = as.character(data$y))

## local membership
loc.grid <- prior(fit, newdata = grid)
contour(seq(-6,6,0.2), seq(-4,4,0.2), matrix(loc.grid[,1], length(seq(-6,6,0.2))), add  = TRUE)
}
\seealso{
Other mixtures svm: \code{\link{FLXMCL-class}},
  \code{\link{FLXPwlda}}

Other mixtures svm: \code{\link{FLXMCL-class}},
  \code{\link{FLXPwlda}}
}

