\docType{class}
\name{FLXPwlda}
\alias{FLXPwlda}
\alias{FLXPwlda-class}
\title{Class "FLXPwlda"}
\description{
  Concomitant model class.
}
\section{Slots}{
  \describe{ \item{\code{name}:}{Character string used in
  print methods.} \item{\code{formula}:}{Formula describing
  the model.} \item{\code{x}:}{Model matrix.}
  \item{\code{fit}:}{Function returning the fitted prior
  probabilities.} \item{\code{refit}:}{Function returning
  the fitted concomitant model.} \item{\code{coef}:}{Matrix
  containing the fitted parameters.}
  \item{\code{df}:}{Function for determining the number of
  degrees of freedom used.} }
}
\seealso{
  \code{\link[flexmix]{FLXP-class}}.
}
\keyword{class}

