\docType{methods}
\name{myfitted}
\alias{myfitted}
\alias{myfitted,flexmix-method}
\title{Extract Fitted Values}
\arguments{
  \item{object}{An object of class "flexmix".}

  \item{drop}{Logical. If \code{TRUE} the function tries to
  simplify the return object by combining lists of length 1
  into matrices.}

  \item{aggregate}{Logical. If \code{TRUE} the fitted
  values for each model aggregated over the components are
  returned.}

  \item{\dots}{Currently unused.}
}
\description{
  Extract fitted values from a flexmix object.
}

