\docType{methods}
\name{mypredict}
\alias{mypredict}
\alias{mypredict,FLXdist-method}
\title{Predict Values}
\arguments{
  \item{object}{An object of class "FLXdist".}

  \item{newdata}{Dataframe containing new data.}

  \item{aggregate}{Logical. If \code{TRUE} the predicted
  values for each model aggregated over the components are
  returned.}

  \item{\dots}{Passed to the \code{predict}-method of the
  model class.}
}
\description{
  Predict values for a FLXdist object.
}

