\name{nnetGradient}
\alias{nnetGradient}
\title{Calculate value of objective function and gradient for a fitted neural network.}
\usage{nnetGradient(object)
}
\description{Calculate value of objective function and gradient for a fitted neural network.}
\arguments{\item{object}{Object inheriting from \code{"nnet"}.}
}

