% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_nnet.R
\docType{data}
\name{nnetModel}
\alias{deviance.nnet}
\alias{estfun.nnet}
\alias{nnetModel}
\alias{predict.nnetModel}
\alias{reweight.nnetModel}
\title{Combine Model-based Recursive Partitioning with Neural Networks}
\format{An object of class \code{StatModel} of length 1.}
\usage{
nnetModel

\\method{reweight}{nnetModel}(object, weights, ...)

\\method{deviance}{nnet}(object, ...)

\\method{estfun}{nnet}(x, ...)

predict.nnetModel(object, out = c("class", "posterior"), newdata, ...)
}
\arguments{
\item{object}{An object of class "nnetModel" and "nnet", respectively.}

\item{weights}{A vector of observation weights.}

\item{x}{An object of class "nnet".}

\item{out}{Should class labels or posterior probabilities be returned?}

\item{newdata}{A \code{data.frame} of cases to be classified.}

\item{\dots}{Further arguments (e.g. to \code{\link[nnet]{nnet}} and \code{\link{nnetRep}}).}
}
\value{
\code{reweight}: The re-weighted fitted "nnetModel" object. \cr
\code{deviance}: The value of the objective function extracted from \code{object}. \cr
\code{estfun}: The empirical estimating (or score) function, i.e. the derivatives of the objective function with respect
  to the parameters, evaluated at the training data. \cr
\code{predict}: Either a vector of predicted class labels or a matrix of class posterior probabilities.
}
\description{
Combine Model-based Recursive Partitioning with Neural Networks.
}
\details{
This page lists all ingredients to combine Neural Networks with Model-Based Recursive Partitioning
(\code{\link[party]{mob}} from package \pkg{party}). See the example for how to do that.

\code{nnetModel} is an object of class \code{\link[modeltools]{StatModel-class}} implemented in package \pkg{modeltools} that
provides an infra-structure for an unfitted \code{\link[nnet]{nnet}} model.

Moreover, methods for \code{\link[nnet]{nnet}} and \code{nnetModel} objects for the generic functions
\code{\link[party]{reweight}}, \code{\link[stats]{deviance}}, \code{\link[sandwich]{estfun}}, and
\code{\link[stats]{predict}} are provided.
}
\examples{
library(benchData)

data <- vData(500)
x <- seq(0,1,0.05)
grid <- expand.grid(x.1 = x, x.2 = x)

fit <- mob(y ~ x.1 + x.2 | x.1 + x.2, data = data, model = nnetModel, size = 1, trace = FALSE,
control = mob_control(objfun = deviance, minsplit = 200))

## predict posterior probabilities
pred <- predict(fit, newdata = grid, out = "posterior")
post <- do.call("rbind", pred)

image(x, x, matrix(as.numeric(post[,1]), length(x)), xlab = "x.1", ylab = "x.2")
contour(x, x, matrix(as.numeric(post[,1]), length(x)), levels = 0.5, add = TRUE)
points(data$x, pch = as.character(data$y))

## predict node membership
splits <- predict(fit, newdata = grid, type = "node")
contour(x, x, matrix(splits, length(x)), levels = min(splits):max(splits), add = TRUE, lty = 2)

## training error
mean(predict(fit) != data$y)

}
\references{
Zeileis, A., Hothorn, T. and Kornik, K. (2008), Model-based recursive partitioning. 
\emph{Journal of Computational and Graphical Statistics}, \bold{17(2)} 492--514.
}
\seealso{
\code{\link[party]{reweight}}, \code{\link[stats]{deviance}}, \code{\link[sandwich]{estfun}}, \code{\link[stats]{predict}}.
}
\keyword{datasets}

