% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnetRep.R
\name{nnetRep}
\alias{nnetRep}
\title{Fit a Neural Network Repeatedly for Different Starting Values}
\usage{
nnetRep(reps = 1, ...)
}
\arguments{
\item{reps}{Number of repetitions. Defaults to 1. \code{reps} larger one does not make sense if \code{Wts} is specified.}

\item{\dots}{Further arguments to \code{\link[nnet]{nnet}}.}
}
\value{
An object of class \code{"nnet"}.
}
\description{
Fit a neural network repeatedly for different starting values.
}
\examples{
fit <- nnetRep(5, Species ~ ., data = iris, size = 2, decay = 0.2, trace = FALSE)
predict(fit)

}
\seealso{
\code{\link[nnet]{nnet}}.
}

