% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oslda.R
\name{oslda}
\alias{oslda}
\alias{oslda.data.frame}
\alias{oslda.default}
\alias{oslda.formula}
\alias{oslda.matrix}
\title{Observation Specific Linear Discriminant Analysis}
\usage{
oslda(x, ...)

\method{oslda}{formula}(formula, data, ..., subset, na.action)

\method{oslda}{data.frame}(x, ...)

\method{oslda}{matrix}(x, grouping, ..., subset, na.action = na.fail)

\method{oslda}{default}(x, grouping, wf = c("biweight", "cauchy", "cosine",
  "epanechnikov", "exponential", "gaussian", "optcosine", "rectangular",
  "triangular"), bw, k, nn.only = TRUE, method = c("unbiased", "ML"), ...)
}
\arguments{
\item{x}{(Required if no \code{formula} is given as principal argument.) A \code{matrix} or \code{data.frame} or \code{Matrix} containing the explanatory variables.}

\item{formula}{A \code{formula} of the form \code{groups ~ x1 + x2 + \dots}, that is, the response
is the grouping \code{factor} and the right hand side specifies the (non-\code{factor})
discriminators.}

\item{data}{A \code{data.frame} from which variables specified in \code{formula} are to be taken.}

\item{subset}{An index vector specifying the cases to be used in the training sample. (NOTE: If given, this argument must be named.)}

\item{na.action}{A function to specify the action to be taken if NAs are found. The default action is first
the \code{na.action} setting of \code{\link{options}} and second \code{\link{na.fail}} if that is unset. 
An alternative is \code{\link{na.omit}}, which leads to rejection of cases with missing values on any required 
variable. (NOTE: If given, this argument must be named.)}

\item{grouping}{(Required if no \code{formula} is given as principal argument.) A \code{factor} specifying
the class membership for each observation.}

\item{wf}{A window function which is used to calculate weights that are introduced into 
the fitting process. Either a character string or a function, e.g. \code{wf = function(x) exp(-x)}.
For details see the documentation for \code{\link[=biweight]{wfs}}.}

\item{bw}{(Required only if \code{wf} is a string.) The bandwidth parameter of the window function. (See \code{\link[=biweight]{wfs}}.)}

\item{k}{(Required only if \code{wf} is a string.) The number of nearest neighbors of the decision boundary to be used in the fitting process. (See \code{\link[=biweight]{wfs}}.)}

\item{nn.only}{(Required only if \code{wf} is a string indicating a window function with infinite support and if \code{k} is specified.) Should
only the \code{k} nearest neighbors or all observations receive positive weights? (See \code{\link[=biweight]{wfs}}.)}

\item{method}{Method for scaling the pooled weighted covariance matrix, either \code{"unbiased"} or maximum-likelihood (\code{"ML"}). Defaults to \code{"unbiased"}.}

\item{\dots}{Further arguments.}
}
\value{
An object of class \code{"oslda"}, a \code{list} containing the following components:
  \item{x}{A \code{matrix} containing the explanatory variables.}
  \item{grouping}{A \code{factor} specifying the class membership for each observation.}
  \item{counts}{The number of observations per class.}
  \item{lev}{The class labels (levels of \code{grouping}).}  
  \item{N}{The number of observations.}
  \item{wf}{The window function used. Always a function, even if the input was a string.}
  \item{bw}{(Only if \code{wf} is a string or was generated by means of one of the functions documented in \code{\link[=biweight]{wfs}}.) 
  The bandwidth used, \code{NULL} if \code{bw} was not specified.}
	 \item{k}{(Only if \code{wf} is a string or was generated by means of one of the functions documented in \code{\link[=biweight]{wfs}}.) 
  The number of nearest neighbors used, \code{NULL} if \code{k} was not specified.}
  \item{nn.only}{(Logical. Only if \code{wf} is a string or was generated by means of one of the functions documented in \code{\link[=biweight]{wfs}} and if \code{k} was
 specified.) \code{TRUE} if only the \code{k} nearest neighbors recieve a positive weight, \code{FALSE} otherwise.}
  \item{adaptive}{(Logical.) \code{TRUE} if the bandwidth of \code{wf} is adaptive to the local density of data points, \code{FALSE} if the bandwidth
  is fixed.}
  \item{method}{The method for scaling the weighted covariance matrices, either \code{"unbiased"} or \code{"ML"}.}
  \item{variant}{(Only if \code{wf} is a string or one of the window functions documented in \code{\link[=biweight]{wfs}} is used, for internal use only). 
  An integer indicating which weighting scheme is implied by \code{bw}, \code{k} and \code{nn.only}.}
  \item{call}{The (matched) function call.}
}
\description{
A localized version of Linear Discriminant Analysis.
}
\details{
This is an alternative implementation of Local Linear Discriminant Analysis proposed by Czogiel et al. (2007) and 
implemented in \code{\link[klaR]{loclda}} in package \pkg{klaR}.

The name of the window function (\code{wf}) can be specified as a character string.
In this case the window function is generated internally in \code{predict.oslda}. Currently
supported are \code{"biweight"}, \code{"cauchy"}, \code{"cosine"}, \code{"epanechnikov"}, 
\code{"exponential"}, \code{"gaussian"}, \code{"optcosine"}, \code{"rectangular"} and 
\code{"triangular"}.

Moreover, it is possible to generate the window functions mentioned above in advance 
(see \code{\link[=biweight]{wfs}}) and pass them to \code{oslda}.

Any other function implementing a window function can also be used as \code{wf} argument.
This allows the user to try own window functions.
See help on \code{\link[=biweight]{wfs}} for details.

If the predictor variables include factors, the formula interface must be used in order 
to get a correct model matrix.
}
\examples{
fit <- oslda(Species ~ Sepal.Length + Sepal.Width, data = iris,
    wf = "gaussian", bw = 0.5, method = "ML")
pred <- predict(fit)
mean(pred$class != iris$Species)

}
\references{
Czogiel, I., Luebke, K., Zentgraf, M. and Weihs, C. (2007), Localized linear discriminant analysis.
In Decker, R. and Lenz, H.-J., editors, Advances in Data Analysis, volume 33 of Studies in Classification,
Data Analysis, and Knowledge Organization, pages 133--140, Springer, Berlin Heidelberg.
}
\seealso{
\code{\link{predict.oslda}}.

Other observation_specific lda: \code{\link{predict.oslda}}

Other observation_specific lda: \code{\link{predict.oslda}}

Other observation_specific lda: \code{\link{predict.oslda}}

Other observation_specific lda: \code{\link{predict.oslda}}

Other observation_specific lda: \code{\link{predict.oslda}}
}
\keyword{classif}
\keyword{multivariate}

