% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmultinom.R
\name{osmultinom}
\alias{osmultinom}
\title{Observation Specific Multinomial Log-linear Models}
\usage{
osmultinom(formula, data, subset, na.action, contrasts = NULL,
  censored = FALSE, model = FALSE, ...)
}
\arguments{
\item{formula}{A formula expression as for regression models, of the form
\code{response ~ predictors}. The response should be a factor or a
 matrix with K columns. If \code{censored = FALSE} this is required to be a 
zero-one indicator matrix.
A log-linear model is fitted, with coefficients zero for the first
class. An offset can be included: it should be a numeric matrix with K columns
if the response is either a matrix with K columns or 
a factor with K > 2
classes, or a numeric vector for a response factor with 2 levels.
See the documentation of \code{\link{formula}()} for other details.}

\item{data}{An optional data frame in which to interpret the variables occurring
in \code{formula}.}

\item{subset}{An index vector specifying the cases to be used in the training sample. (NOTE: If given, this argument must be named.)}

\item{na.action}{A function to specify the action to be taken if \code{NA}s are found. The default action is for the
procedure to fail. 
An alternative is \code{\link{na.omit}}, which leads to rejection of cases with missing values on any required 
variable. (NOTE: If given, this argument must be named.)}

\item{contrasts}{A list of contrasts to be used for some or all of the factors appearing as variables in the model formula.}

\item{censored}{Logical. If the response is a matrix with \eqn{K > 2} classes, interpret the entries as one for possible classes, zero 
for impossible classes. Defaults to \code{FALSE}.}

\item{model}{Logical. If \code{TRUE}, the model frame is saved as component \code{model} of the returned object.}

\item{\dots}{Additional arguments for \code{\link{osnnet}}, including the window function and bandwidth
 parameters used to generate observation weights:
\describe{
    \item{\code{wf}}{A window function which is used to calculate weights that are introduced into 
        the fitting process. Either a character string or a function, e.g. \code{wf = function(x) exp(-x)}.
        For details see the documentation for \code{\link[=biweight]{wfs}}.}
    \item{\code{bw}}{(Required only if \code{wf} is a string.) The bandwidth parameter of the window function. (See \code{\link[=biweight]{wfs}}.)}
 \item{\code{k}}{(Required only if \code{wf} is a string.) The number of nearest neighbors of the decision boundary to be used in the fitting process. 
    (See \code{\link[=biweight]{wfs}}.)}
    \item{\code{nn.only}}{(Required only if \code{wf} is a string indicating a window function with infinite support and if \code{k} is specified.) Should
        only the \code{k} nearest neighbors or all observations receive positive weights? (See \code{\link[=biweight]{wfs}}.)}
    \item{\code{itr}}{Number of iterations for model fitting, defaults to 3. See also the Details section.}
}}
}
\value{
An object of class \code{osmultinom}, inheriting from \code{"osnnet"}, a \code{list} containing the following components:
  \item{x}{A \code{matrix} containing the explanatory variables.}
  \item{y}{If argument \code{y} was a factor a class inicator matrix, otherwise a \code{matrix} that contains \code{y} unchanged.}
  \item{...}{}
  \item{mask}{The \code{mask} vector used.}
  \item{maxit}{The \code{maxit} argument used.}
  \item{trace}{The \code{trace} argument used.}
  \item{abstol}{The \code{abstol} argument used.}
  \item{reltol}{The \code{reltol} argument used.}
  \item{lev}{If \code{y} is a factor the class labels (levels of \code{y}).}  
  \item{wf}{The window function used. Always a function, even if the input was a string.}
  \item{bw}{(Only if \code{wf} is a string or was generated by means of one of the functions documented in \code{\link[=biweight]{wfs}}.) 
  The bandwidth used, \code{NULL} if \code{bw} was not specified.}
	 \item{k}{(Only if \code{wf} is a string or was generated by means of one of the functions documented in \code{\link[=biweight]{wfs}}.) 
  The number of nearest neighbors used, \code{NULL} if \code{k} was not specified.}
  \item{nn.only}{(Logical. Only if \code{wf} is a string or was generated by means of one of the functions documented in \code{\link[=biweight]{wfs}} and if \code{k} was
 specified.) \code{TRUE} if only the \code{k} nearest neighbors recieve a positive weight, \code{FALSE} otherwise.}
  \item{adaptive}{(Logical.) \code{TRUE} if the bandwidth of \code{wf} is adaptive to the local density of data points, \code{FALSE} if the bandwidth
  is fixed.}
  \item{variant}{(Only if \code{wf} is a string or one of the window functions documented in \code{\link[=biweight]{wfs}} is used, for internal use only). 
  An integer indicating which weighting scheme is implied by \code{bw}, \code{k} and \code{nn.only}.}
  \item{call}{The (matched) function call.}
}
\description{
A localized version of multinomial regression.
}
\details{
This is a localized version of multinomial regression where a multinomial regression model is fitted for 
each test observation based on the training data near the trial point. It
is based on the function \code{\link[nnet]{multinom}} from package \pkg{nnet}.
\code{\link{osnnet}} is called internally.

In \code{osmultinom} no fitting is done. In the prediction step an individual model for every test
observation is fitted.
Observation weights that reflect the importance of training observations for the fit
at a particular test observation are calculated internally in \code{\link{predict.osmultinom}}.
For this reason not all types of response in \code{formula} are allowed and \code{osmultinom} does not take all arguments 
that can be passed to \code{\link[nnet]{multinom}}.
As response in \code{formula} factors and matrices are allowed. If \code{censored = FALSE} only zero-one class indicator matrices are 
allowed.
Argument \code{weights} is missing since observation weights are calculated internally in \code{\link{predict.osmultinom}}.
\code{summ} that specifies a method to summarize rows of the model matrix is missing since this requires adjustment of the case weights.
Also \code{Hess} is not supported.
}
\examples{
fit <- osmultinom(Species ~ Sepal.Length + Sepal.Width, data = iris,
    wf = "gaussian", bw = 0.5)
pred <- predict(fit)
mean(pred != iris$Species)

}
\references{
Czogiel, I., Luebke, K., Zentgraf, M. and Weihs, C. (2007), Localized linear discriminant analysis.
In Decker, R. and Lenz, H.-J., editors, Advances in Data Analysis, volume 33 of Studies in Classification,
Data Analysis, and Knowledge Organization, pages 133--140, Springer, Berlin Heidelberg.

Ripley, B. D. (1996) \emph{Pattern Recognition and Neural Networks}. Cambridge.

Venables, W. N. and Ripley, B. D. (2002) \emph{Modern Applied Statistics with S}. Fourth edition. Springer.
}
\seealso{
\code{\link{predict.osmultinom}}, \code{\link{osnnet}}, \code{\link[nnet]{nnet}}.

Other observation_specific multinom: \code{\link{predict.osmultinom}}
}
\keyword{neural}

