% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daqda.R
\name{predict.daqda}
\alias{predict.daqda}
\title{Classify Multivariate Observations Based on Discriminant Adaptive Quadratic Discriminant Analysis}
\usage{
predict.daqda(object, newdata, ...)
}
\arguments{
\item{object}{Object of class \code{"daqda"}.}

\item{newdata}{A \code{data.frame} of cases to be classified or, if \code{object} has a
\code{formula}, a \code{data.frame} with columns of the same names as the
variables used. A vector will be interpreted as a row
vector. If \code{newdata} is missing, an attempt will be made to
retrieve the data used to fit the \code{daqda} object.}

\item{\dots}{Further arguments.}
}
\value{
A \code{list} with components:
\item{class}{The predicted class labels (a \code{factor}).}
\item{posterior}{Matrix of class posterior probabilities.}
}
\description{
Classify multivariate observations in conjunction with \code{\link{daqda}}.
}
\details{
This function is a method for the generic function \code{predict()} for class 
\code{"daqda"}. 
It can be invoked by calling \code{predict(x)} for an object \code{x} of the 
appropriate class, or directly by calling \code{predict.daqda(x)} regardless of 
the class of the object.
}
\examples{
## comparison with qda:
library(MASS)
fit1 <- qda(Species ~ Sepal.Length + Sepal.Width, data = iris)
pred <- predict(fit1)
mean(pred$class != iris$Species)

fit2 <- daqda(Species ~ Sepal.Length + Sepal.Width, data = iris,
    wf = "gaussian", bw = 0.5)
pred <- predict(fit2)
mean(pred$class != iris$Species)

## plot of decision boundary (maximum posterior probabilities):
grid <- expand.grid(Sepal.Length = seq(4,8,0.1), Sepal.Width = seq(2,5,0.1)) 

predgrid1 <- predict(fit1, newdata = grid)$posterior
predgrid2 <- predict(fit2, newdata = grid)$posterior

par(mfrow = c(1,2))
contour(seq(4,8,0.1), seq(2,5,0.1), 
    matrix(as.numeric(apply(predgrid1, 1, max)), nrow = length(seq(4,8,0.1))))
contour(seq(4,8,0.1), seq(2,5,0.1), 
    matrix(as.numeric(apply(predgrid2, 1, max)), nrow = length(seq(4,8,0.1))))
points(iris$Sepal.Length, iris$Sepal.Width, pch = 19, 
    cex = fit2$weights[[3]]*2, col = as.numeric(iris$Species))

}
\references{
Hand, D. J., Vinciotti, V. (2003), Local versus global models for classification problems: 
Fitting models where it matters, \emph{The American Statistician}, \bold{57(2)} 124--130.
}
\seealso{
\code{\link{daqda}}, \code{\link{wqda}}, \code{\link{dalr}}.
}
\keyword{classif}
\keyword{multivariate}

