% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majorityGenerative.R
\name{predict.majorityGenerative}
\alias{predict.majorityGenerative}
\title{Classify Observations with the Weighted Majority Classifier}
\usage{
predict.majorityGenerative(object, newdata, prior = object$prior, ...)
}
\arguments{
\item{object}{Object of class \code{"majorityGenerative"}.}

\item{newdata}{A \code{data.frame} of cases to be classified or, if \code{object} has a
\code{formula}, a \code{data.frame} with columns of the same names as the
variables used.  A vector will be interpreted as a row
vector.  If \code{newdata} is missing, an attempt will be made to
retrieve the data used to fit the \code{majorityGenerative} object.}

\item{prior}{The class prior probabilities. By default the proportions in the training data set.}

\item{\dots}{Further arguments.}
}
\value{
A \code{list} with components:
\item{class}{The predicted class labels (a \code{factor}).}
\item{posterior}{Matrix of class posterior probabilities.}
}
\description{
Classify observations in conjunction with \code{\link{majorityGenerative}}.
}
\details{
This function is a method for the generic function \code{predict()} for class
 \code{"majorityGenerative"}.
 It can be invoked by calling \code{predict(x)} for an object \code{x} of the
 appropriate class, or directly by calling \code{predict.majorityGenerative(x)} regardless of
 the class of the object.
}
\examples{
library(mlbench)
data(PimaIndiansDiabetes)

train <- sample(nrow(PimaIndiansDiabetes), 500)

# weighting observations from classes pos and neg according to their
# frequency in the data set:
ws <- as.numeric(1/table(PimaIndiansDiabetes$diabetes)
    [PimaIndiansDiabetes$diabetes])

fit <- majorityGenerative(diabetes ~ ., data = PimaIndiansDiabetes, weights = ws,
    subset = train)
pred <- predict(fit, newdata = PimaIndiansDiabetes[-train,])
mean(pred$class != PimaIndiansDiabetes$diabetes[-train])
}
\seealso{
Other majority: \code{\link{majorityGenerative}},
  \code{\link{majority}}, \code{\link{predict.majority}}
}

