% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oslda.R
\name{predict.oslda}
\alias{predict.oslda}
\title{Classify Multivariate Observations Based on Observation Specific Linear Discriminant Analysis}
\usage{
predict.oslda(object, newdata, ...)
}
\arguments{
\item{object}{Object of class \code{"oslda"}.}

\item{newdata}{A \code{data.frame} of cases to be classified or, if \code{object} has a
\code{formula}, a \code{data.frame} with columns of the same names as the
variables used. A vector will be interpreted as a row
vector. If \code{newdata} is missing, an attempt will be made to
retrieve the data used to fit the \code{oslda} object.}

\item{\dots}{Further arguments.}
}
\value{
A \code{list} with components:
\item{class}{The predicted class labels (a \code{factor}).}
\item{posteriors}{Matrix of class posterior probabilities.}
}
\description{
Classify multivariate observations in conjunction with \code{\link{oslda}}.
}
\details{
This function is a method for the generic function \code{predict()} for class 
\code{"oslda"}. 
It can be invoked by calling \code{predict(x)} for an object \code{x} of the 
appropriate class, or directly by calling \code{predict.oslda(x)} regardless of 
the class of the object.
}
\examples{
      fit <- oslda(Species ~ Sepal.Length + Sepal.Width, data = iris,
                   wf = "gaussian", bw = 0.5, method = "ML")
     pred <- predict(fit)
     mean(pred$class != iris$Species)

}
\seealso{
\code{\link{oslda}}.

Other observation_specific lda: \code{\link{oslda}}
}
\keyword{classif}

