% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ossvm.R
\name{predict.ossvm}
\alias{predict.ossvm}
\title{Predict New Observations with Observation Specific Support Vector Machines}
\usage{
predict.ossvm(object, newdata, decision.values = FALSE, probability = FALSE,
  ..., na.action = na.omit)
}
\arguments{
\item{object}{Object of class \code{"ossvm"}.}

\item{newdata}{An object containing the new input data: either a matrix or a sparse matrix 
(object of class \code{\link[Matrix]{Matrix}} provided by the \pkg{Matrix} package, or of class 
\code{\link[SparseM]{matrix.csr}} provided by the \pkg{SparseM} package, or of class 
\code{\link[slam]{simple_triplet_matrix}} provided by the \pkg{slam} package). A vector will be 
transformed to a (n x 1) matrix.}

\item{decision.values}{Logical controlling whether the decision values of all binary classifiers 
computed in multiclass classification shall be computed and returned.}

\item{probability}{Logical indicating whether class probabilities should be computed and returned.}

\item{na.action}{A function to specify the action to be taken if 'NA's are found. The default 
action is na.omit, which leads to rejection of cases with missing 
  values on any required variable. An alternative is na.fail, which causes an error if NA cases are found. 
(NOTE: If given, this argument must be named.)}

\item{\dots}{Further arguments.}
}
\value{
A vector of predicted values (for classification: a vector of labels, for density 
estimation: a logical vector). If \code{decision.value} is \code{TRUE}, 
the vector gets a \code{"decision.values"} attribute containing a \eqn{n * c} matrix (\eqn{n} number of 
predicted values, \eqn{c} number of classifiers) of all \eqn{c} binary 
classifiers' decision values. There are \eqn{k * (k - 1) / 2} classifiers (\eqn{k} number of classes). 
The colnames of the matrix indicate the labels of the two classes. 
If \code{probability} is \code{TRUE}, the vector gets a \code{"probabilities"} attribute 
containing a \eqn{n * k} matrix of the class probabilities.
}
\description{
Classify multivariate observations in conjunction with \code{\link{ossvm}}.
}
\details{
This function is a method for the generic function \code{predict()} for class 
\code{"ossvm"}. 
It can be invoked by calling \code{predict(x)} for an object \code{x} of the 
appropriate class, or directly by calling \code{predict.ossvm(x)} regardless of 
the class of the object.
}
\seealso{
\code{\link{ossvm}}.

Other observation_specific svm: \code{\link{ossvm}}
}
\keyword{classif}

