% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_nnet.R
\name{root.matrix}
\alias{root.matrix}
\title{Root of a Matrix}
\usage{
root.matrix(X)
}
\arguments{
\item{X}{A symmetric and positive semidefinite matrix.}
}
\value{
A symmetric matrix of same dimension as \code{X}.
}
\description{
Computes the root of a symmetric and positive semidefinite matrix.
}
\details{
Eigenvalues smaller than zero are replaced by the minimum of the smallest positive eigenvalue and their absolute value.
}

