% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLXMCL.R, R/FLXMCLlda.R, R/FLXMCLmajority.R, R/FLXMCLmajorityGenerative.R, R/FLXMCLmultinom.R, R/FLXMCLnnet.R, R/FLXMCLqda.R, R/FLXMCLsvm.R
\docType{class}
\name{FLXMCL-class}
\alias{FLXMCL-class}
\alias{FLXMCLlda-class}
\alias{FLXMCLmajority-class}
\alias{FLXMCLmajorityGenerative-class}
\alias{FLXMCLmultinom-class}
\alias{FLXMCLnnet-class}
\alias{FLXMCLqda-class}
\alias{FLXMCLsvm-class}
\title{Class "FLXMCL"}
\description{
Model specification for mixtures of classifiers.
}
\details{
Objects can be created by calls of the form \code{new("FLXMCL", ...)},
  typically inside driver functions like \code{\link{FLXMCLlda}} or
  \code{\link{FLXMCLmultinom}}. 
Class FLXMCL extends class \code{\link[flexmix]{FLXM-class}} from package \pkg{flexmix} directly
and has exactly the same slots.
}
\seealso{
\code{\link[flexmix]{FLXM-class}}.

Other mixtures: \code{\link{FLXPwlda-class}},
  \code{\link{FLXPwlda}}, \code{\link{myStepFlexmix}},
  \code{\link{myfitted}}, \code{\link{mypredict}}

Other mixtures lda: \code{\link{FLXMCLlda}}

Other mixtures majority: \code{\link{FLXMCLmajorityGenerative}},
  \code{\link{FLXMCLmajority}}

Other mixtures majority: \code{\link{FLXMCLmajorityGenerative}},
  \code{\link{FLXMCLmajority}}

Other mixtures multinom: \code{\link{FLXMCLmultinom}}

Other mixtures nnet: \code{\link{FLXMCLnnet}}

Other mixtures qda: \code{\link{FLXMCLqda}}

Other mixtures svm: \code{\link{FLXMCLsvm}},
  \code{\link{FLXPwlda}}
}
\keyword{classes}

