% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLXPwlda.R
\docType{class}
\name{FLXPwlda-class}
\alias{FLXPwlda-class}
\title{Classes "FLXPwlda" and "FLXPwqda"}
\description{
Concomitant model class.
}
\seealso{
\code{\link[flexmix]{FLXP-class}}.

Other mixtures: \code{\link{FLXMCL-class}},
  \code{\link{FLXPwlda}}, \code{\link{myStepFlexmix}},
  \code{\link{myfitted}}, \code{\link{mypredict}}
}
\keyword{class}

