% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLXPwlda.R
\name{FLXPwlda}
\alias{FLXPwlda}
\alias{FLXPwqda}
\title{Creator Functions for the Concomitant Variable Model based on Linear or Quadratic Discriminant Analysis}
\usage{
FLXPwlda(formula = ~.)

FLXPwqda(formula = ~.)
}
\arguments{
\item{formula}{A formula for determining the model matrix of the concomitant variables.}
}
\value{
Object of class \code{FLXPwlda} or \code{FLXPwqda} which both extend class \code{FLXP} directly and are used for method dispatching.
}
\description{
Creator function for the concomitant variable model. Priors are modeled by Linear or Quadratic Discriminant Analysis.
}
\examples{
library(benchData)
data <- flashData(1000)
x1 <- seq(-6,6,0.2)
x2 <- seq(-4,4,0.2)
grid <- expand.grid(x.1 = x1, x.2 = x2)

cluster <- kmeans(data$x, center = 2)$cluster
model <- FLXMCLmultinom(trace = FALSE)
fit <- flexmix(y ~ x.1 + x.2, data = as.data.frame(data), concomitant = FLXPwlda(~ x.1 + x.2), model = model, cluster = cluster, control = list(verb = 1))

## prediction for single component models without aggregation
pred.grid <- predict(fit, newdata = grid)
image(x1, x2, matrix(pred.grid[[1]][,1], length(x1)))
contour(x1, x2, matrix(pred.grid[[1]][,1], length(x1)), add = TRUE)
points(data$x, pch = as.character(data$y))

image(x1, x2, matrix(pred.grid[[2]][,1], length(x1)))
contour(x1, x2, matrix(pred.grid[[2]][,1], length(x1)), add = TRUE)
points(data$x, pch = as.character(data$y))

## prediction with aggregation depending on membership in mixture components
pred.grid <- mypredict(fit, newdata = grid, aggregate = TRUE)
image(x1, x2, matrix(pred.grid[[1]][,1], length(x1)))
contour(x1, x2, matrix(pred.grid[[1]][,1], length(x1)), add  = TRUE)
points(data$x, pch = as.character(data$y))

## local membership
loc.grid <- prior(fit, newdata = grid)
contour(x1, x2, matrix(loc.grid[,1], length(x1)), add  = TRUE)
}
\seealso{
\code{\link[flexmix]{FLXPmultinom}}.

Other mixtures: \code{\link{FLXMCL-class}},
  \code{\link{FLXPwlda-class}},
  \code{\link{myStepFlexmix}}, \code{\link{myfitted}},
  \code{\link{mypredict}}

Other mixtures svm: \code{\link{FLXMCL-class}},
  \code{\link{FLXMCLsvm}}
}

