% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLXMCL.R
\docType{methods}
\name{myfitted}
\alias{myfitted}
\alias{myfitted,flexmix-method}
\title{Extract Fitted Values}
\usage{
myfitted(object, drop = TRUE, aggregate = FALSE, ...)

\S4method{myfitted}{flexmix}(object, drop = TRUE, aggregate = FALSE, ...)
}
\arguments{
\item{object}{An object of class "flexmix".}

\item{drop}{Logical. If \code{TRUE} the function tries to
simplify the return object by combining lists of length 1 into
matrices.}

\item{aggregate}{Logical. If \code{TRUE} the fitted values for
each model aggregated over the components are returned.}

\item{\dots}{Currently unused.}
}
\description{
Extract fitted values from a flexmix object.
}
\seealso{
Other mixtures: \code{\link{FLXMCL-class}},
  \code{\link{FLXPwlda-class}}, \code{\link{FLXPwlda}},
  \code{\link{myStepFlexmix}}, \code{\link{mypredict}}

Other mixtures: \code{\link{FLXMCL-class}},
  \code{\link{FLXPwlda-class}}, \code{\link{FLXPwlda}},
  \code{\link{myStepFlexmix}}, \code{\link{mypredict}}
}

