% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLXMCL.R
\docType{methods}
\name{mypredict}
\alias{mypredict}
\alias{mypredict,FLXdist-method}
\title{Predict Values}
\usage{
mypredict(object, newdata = list(), aggregate = FALSE, ...)

\S4method{mypredict}{FLXdist}(object, newdata = list(), aggregate = FALSE,
  ...)
}
\arguments{
\item{object}{An object of class "FLXdist".}

\item{newdata}{Dataframe containing new data.}

\item{aggregate}{Logical. If \code{TRUE} the predicted values for
each model aggregated over the components are returned.}

\item{\dots}{Passed to the \code{predict}-method of the model class.}
}
\description{
Predict values for a FLXdist object.
}
\seealso{
Other mixtures: \code{\link{FLXMCL-class}},
  \code{\link{FLXPwlda-class}}, \code{\link{FLXPwlda}},
  \code{\link{myStepFlexmix}}, \code{\link{myfitted}}

Other mixtures: \code{\link{FLXMCL-class}},
  \code{\link{FLXPwlda-class}}, \code{\link{FLXPwlda}},
  \code{\link{myStepFlexmix}}, \code{\link{myfitted}}
}

