% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmultinom.R
\name{predict.osmultinom}
\alias{predict.osmultinom}
\title{Predict New Examples Based on Observation Specific Multinomial Log-linear Models}
\usage{
predict.osmultinom(object, newdata, type = c("class", "probs"), ...)
}
\arguments{
\item{object}{Object of class \code{"osmultinom"}.}

\item{newdata}{A \code{data.frame} of cases to be classified or, if \code{object} has a
\code{formula}, a \code{data.frame} with columns of the same names as the
variables used. A vector will be interpreted as a row
vector. If \code{newdata} is missing, an attempt will be made to
retrieve the data used in \code{\link{osmultinom}}.}

\item{type}{Type of output.}

\item{\dots}{Arguments passed to or from other methods.}
}
\value{
If \code{type = "class"}, a \code{factor} containing the predicted class labels;
  if \code{type = "probs"}, a \code{matrix} of class posterior probabilities. In case of a binary classification
 problem the matrix contains only the posteriors for the second factor level.
}
\description{
Predict new examples in conjunction with \code{\link{osmultinom}}.
}
\details{
This function is a method for the generic function \code{predict()} for class 
\code{"osmultinom"}. 
It can be invoked by calling \code{predict(x)} for an object \code{x} of the 
appropriate class, or directly by calling \code{predict.osmultinom(x)} regardless of 
the class of the object.
}
\examples{
    fit <- osmultinom(Species ~ Sepal.Length + Sepal.Width, data = iris,
                      wf = "gaussian", bw = 0.5)
    pred <- predict(fit)
    mean(pred != iris$Species)

}
\seealso{
\code{\link{osmultinom}}, \code{\link{osnnet}}, \code{\link{predict.osnnet}}, \code{\link[nnet]{predict.nnet}}, \code{\link[nnet]{nnet}}.

Other observation_specific multinom: \code{\link{osmultinom}}
}
\keyword{neural}

